requirements.txt
anthropic>=0.41.0
openai>=1.59.5
google-generativeai==0.8.3
langchain==0.3.13
langchain-community==0.3.13
langchain-chroma==0.1.4
chromadb==0.5.23
sentence-transformers==3.3.1
pymupdf==1.25.2
requests==2.32.3
schedule==1.2.2
python-dotenv==1.0.1
pytz==2024.2
